+++ greenfield/eldenring/data.py
@@ -5557,4 +5557,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:6f17d305d2e33c25b5870436f865914a4bfaaf8bea6b5a5d3d4658023ee762f6', 'module': 'data.py', 'body_sha256': 'sha256:1add59b7ff6b743afac5eefd8b0d64b49dd634bea49a354392c4ee39146526a6'}
+_GEN_STAMP = {'inputs_hash': 'sha256:0a4e0375d8e814753b712888846c4152b4f4371c9a4a87d7f871c70ad7287362', 'module': 'data.py', 'body_sha256': 'sha256:1add59b7ff6b743afac5eefd8b0d64b49dd634bea49a354392c4ee39146526a6'}